Drawerで画像の一部だけ拡大表示できるようにする関数を追加 レーザーの当たり判定を修正

+++ Mirror/GameClient/Lazer.h
@@ -23,8 +23,8 @@ private:
 		float angle;
 	};
 	struct ImageSize {
-		short int width;
-		short int height;
+		float width;
+		float height;
 	};
 public:
 	Lazer( GlobalDataPtr data );
